secure_file_sharing/project/requirements.txt
Flask==2.3.2
cryptography==41.0.1
python-dotenv==1.0.0
